# server/src/Tgm.Roborally.Server/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "contact" : {
      "email" : "nbrugger@student.tgm.ac.at",
      "name" : "Nils Brugger"
    },
    "description" : "This api controlls the flow of a game and provides it's data.\nIt is desiged to be RESTfull so the structure works simmilar as file system.\nThe service will run and only work in a local network, `game.host` is the IP of the Computer hosting the game and will be found via a IP scan",
    "license" : {
      "name" : "Apache 2.0"
    },
    "title" : "Robot Rally Game logic engine",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "description" : "Production",
    "url" : "http://game.host/v1"
  }, {
    "description" : "Debug",
    "url" : "http://localhost:5050/v1"
  } ],
...
  "paths" : {
...
    "/games/{game_id}/entitys/robots/{robot_id}/registers" : {
      "delete" : {
        "description" : "Emptys all register of the robot",
        "operationId" : "clearRegisters",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "robot_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/EntityID"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "clear registers",
        "tags" : [ "Robots" ]
      },
      "get" : {
        "description" : "Returns the content of the registers",
        "operationId" : "getRegisters",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "robot_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/EntityID"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/RobotCommand"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Get register information",
        "tags" : [ "Robots" ]
      },
      "patch" : {
        "description" : "Updates the card in a register",
        "operationId" : "updateRegister",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "robot_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/EntityID"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "change register content",
        "tags" : [ "Robots" ]
      }
    },
    "/games/{game_id}/entitys/robots/{robot_id}/upgrades" : {
      "delete" : {
        "description" : "Removes all upgrades from a robot",
        "operationId" : "clearUpgrades",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "robot_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/EntityID"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Clear Robot Upgrades",
        "tags" : [ "Robots" ]
      },
      "get" : {
        "description" : "Returns the upgrades installed on the robot",
        "operationId" : "getInstalledUpgrades",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "robot_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/EntityID"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/UpgradeID"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Get Robot Upgrades",
        "tags" : [ "Robots" ]
      },
      "patch" : {
        "description" : "Removes a paticular upgrade from the robot",
        "operationId" : "removeRobotUpgrade",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "robot_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/EntityID"
          },
          "style" : "simple"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "upgrade",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/UpgradeID"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Remove Upgrade",
        "tags" : [ "Robots" ]
      }
    },
...
    "/games/{game_id}/entitys/robots/{robot_id}/info" : {
      "get" : {
        "description" : "Returns the status and info about the robot",
        "operationId" : "getRobotStats",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "robot_id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RobotInfo"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Get Robot Informations",
        "tags" : [ "Robots" ]
      }
    },
    "/games/{game_id}/entitys/robots/{robot_id}/actions/stack/" : {
      "get" : {
        "description" : "Retuns a list od all actions in the stack",
        "operationId" : "getActionStack",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "robot_id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/EntityAction"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Get Robot action stack",
        "tags" : [ "Robots" ]
      },
      "post" : {
        "description" : "Adds an action to the stack. The stack is processed using *first in first out*",
        "operationId" : "doEntityAction",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "robot_id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EntityAction"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Add Entity Action to stack",
        "tags" : [ "Robots" ]
      }
    },
    "/games/{game_id}/entitys/robots/{robot_id}/actions/avinable" : {
      "get" : {
        "description" : "Returns a list of EntityActionTypes which are possible to be executed at the moment.",
        "operationId" : "getPossibleActions",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "robot_id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/EntityEventOportunity"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Get Aviable actions",
        "tags" : [ "Robots" ]
      }
    },
    "/games/{game_id}/entitys/robots/" : {
      "get" : {
        "description" : "Returns a list of all robot IDs in this game.\n> A robot is not a player as a player is able to controll multiple robots",
        "operationId" : "getRobots",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/EntityID"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Get all robots",
        "tags" : [ "Robots" ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "GameID" : {
        "description" : "**Unique**<br>\nThis is the parameter a game is identified by",
        "example" : 537,
        "format" : "int32",
        "maximum" : 2048,
        "minimum" : 0,
        "title" : "GameID",
        "type" : "integer"
      },
...
      "Error-Message" : {
        "description" : "A message about that went wrong. Usefull to display users a short and usefull prompt",
        "properties" : {
          "message" : {
            "description" : "A short message describing what happened in human words",
            "type" : "string"
          },
          "error" : {
            "description" : "The error/exception",
            "type" : "string"
          }
        },
        "required" : [ "message" ],
        "title" : "Error-Message",
        "type" : "object"
      },
...
      "EntityID" : {
        "description" : "The unique identification of this entity. <br>\n*!!!* This is not the id of the player<br>\nThis value will be autogenerated by the api and is read only",
        "format" : "int32",
        "minimum" : 0,
        "title" : "EntityID",
        "type" : "integer"
      },
...
      "UpgradeID" : {
        "description" : "The id of an upgrade. **Unique**",
        "maximum" : 10000,
        "minimum" : 0,
        "title" : "UpgradeID",
        "type" : "integer"
      },
...
      "RobotInfo" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Entity"
        }, {
          "$ref" : "#/components/schemas/RobotInfo_allOf"
        } ],
        "description" : "Describes the state/stats of a robot",
        "title" : "RobotInfo"
      },
...
    },
    "securitySchemes" : {
      "Host-token-access" : {
        "description" : "The Host of the game (and of the API) has an unique identificaton granting him game manipulating abilities. By this means it is also possible to have a host and a client at the same machine. \n> This could be used for AI`s (bot players)",
        "in" : "query",
        "name" : "hid",
        "type" : "apiKey"
      },
      "Player-Access-Token" : {
        "description" : "Players need a token to join a lobby/game. If the host sets a password the player needs the password to get the token",
        "in" : "query",
        "name" : "pat",
        "type" : "apiKey"
      }
    }
  }
}